requirements.txt
# runtime libraries
datetime
pandas
python-dateutil
requests

# development libraries
pytest